+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 
 ## [Unreleased]
 
+- refacto download url for peertube runners
 - demo app now create files in storage directory
 - add django_peertube_runner_connector app
 - add basic quality tools + testing suite